+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 _The format is based on [Keep a Changelog](https://https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/)._
 
+## Unreleased
+
+Date: 2020-12-23
+Authors: Erik Vlasblom
+Added:
+
+- Spotify API
+  - Get the top artists and tracks.
+
 ## 1.0.0
 
 Date: 2020-12-05